requirements.txt
numpy>=1.19.4
ninja>=1.10.0
opencv-python>=4.4.0.46
numba>=0.52.0
progress>=1.5
matplotlib>=3.3.3
easydict>=1.9
scipy>=1.5.4
pycocotools>=2.0.2
scikit-learn>=0.20.4
imagesize>=1.2.0
munch>=2.5.0
yapf>=0.30.0
tqdm>=4.55.0
tensorboardX>=2.1
pyyaml
boto3
futures
prettytable
nvsmi
nuscenes-devkit==1.1.7
omegaconf==2.1.1
diskcache>=4.1.0
tenacity>=6.0.0
xarray
transforms3d
pybind11
open3d-python
pytorch-lightning
easydict
